Move getSortedEvents to store utils, clean up lint issues
