+++ src/frontend/calendar/utils.js
@@ -1,8 +1,3 @@
-/**
- * WordPress dependencies
- */
-import { format } from '@wordpress/date';
-
 // Default value for days not in this month.
 const emptyDate = {
 	blank: true,
@@ -45,22 +40,4 @@ export function getRows( year, month ) {
 	}
 
 	return rows;
-}
-
-/**
- * Get the list of events in day-buckets.
- *
- * @param {Array} events
- */
-export function getSortedEvents( events ) {
-	const sortedEvents = {};
-	events.forEach( ( event ) => {
-		const key = format( 'Y-m-d', event.date );
-		if ( sortedEvents.hasOwnProperty( key ) ) {
-			sortedEvents[ key ].push( event );
-		} else {
-			sortedEvents[ key ] = [ event ];
-		}
-	} );
-	return sortedEvents;
 }
